move vanilla todomvc example to create-react-app
