+++ examples/react/.gitignore
@@ -1,12 +1,17 @@
-node_modules/director
-!node_modules/director/build/director.js
+# See https://help.github.com/ignore-files/ for more about ignoring files.
 
-node_modules/react
-!node_modules/react/dist/react-with-addons.js
-!node_modules/react/dist/JSXTransformer.js
-node_modules/todomvc-app-css
-!node_modules/todomvc-app-css/index.css
+# dependencies
+/node_modules
 
-node_modules/todomvc-common
-!node_modules/todomvc-common/base.css
-!node_modules/todomvc-common/base.js
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
